media-automation/requirements.txt
beautifulsoup4==4.12.3
selenium==4.18.1
requests==2.31.0
supabase==2.3.4
python-dotenv==1.0.0
Pillow==10.2.0
lxml==5.1.0
